merge r4345 from branch 2.0 to trunk
feature 1051: ability to add/update a file for an existing photo. For example,
you can add the "high" later. Another example is to update the "web resized"
file (new dimensions is a common example). It also works for thumbnails.
Updating an existing file has no impact on the logical level (list of tags,
list of categories, title, description and so on).


git-svn-id: http://piwigo.org/svn/trunk@4348 68402e56-0260-453c-a942-63ccdbb3a9ee