merge r11040 from branch 2.2 to trunk
bug 1786 fixed: ability to prefilter the synchronize screen with a specific
album by clicking on the new "Synchronize" icon on album admin page or album
admin list.

This is not the solution proposed by Gotcha (ie a treeview to select the album
on the synchronize screen) but it solves the same problem.



git-svn-id: http://piwigo.org/svn/trunk@11041 68402e56-0260-453c-a942-63ccdbb3a9ee