feature 2352 added: pwg.categories.getList now returns a tn_url parameter. This is the album thumbnail.
Warning: if the API method is called with $params['public'], the album
thumbnail may be not accurate. The thumbnail can be viewed by the connected
user, but maybe not by the guest. Changing the filtering method would be too
complicated for now. We will simply avoid to persist the
user_representative_picture_id in the database if $params['public']


git-svn-id: http://piwigo.org/svn/branches/2.3@12537 68402e56-0260-453c-a942-63ccdbb3a9ee