feature #392, authentication keys, history log
When a user successfully performs an authentication with an auth_key, Piwigo
registers it in the history table.

For now, it is not shown/searchable in the history screen, but we can add it
in the future and we can provide a plugin with specific details about
authentication keys usage.