feature #478, improve performances
By reducing frequency of update on user_infos.last_visit. Instead of updating
on each page visit, we wait at least a session length, which is 30 minutes by
default.

@modus75 suggested to use function log_user instead, which is a better design.
Unfortunately it would lead to much less accuracy, because on many systems PHP
sessions are never purged (Debian with default settings). It would be possible
to come back everyday and never update the last_visit, as long as the web
browser is not closed.