fixes #556, tag manager actions in a list, not buttons
The actions are no longer displayed a individual submit buttons, but as a
select, like on other managers (users, photos, groups).

The edit/duplicate/merge forms are still displayed only after a first submit.
To be improved to be available right after the action selection.

When we display an action form edit/duplicate/merge, we don't show other
fieldsets on the page to focus on the important action.