Static analysis support with Phan This rather scary-looking patch adds a .phan/config.php file along with some stubs and a bunch of fixes to make it possible to use Phan to catch bugs.
To use it install phan somewhere with 'composer require phan/phan' and
then from the top Piwigo directory run: <path>/vendor/bin/phan -p
With this patch there are a little under 900 issues to fix. Many of
them are false positives due to global scope code and variables.
With some suppresses and better phpdoc typing, it should be possible
to get it to 0. You can see the type of fixes it takes in this patch.

And yes, it is a scary-looking PR. Feel free to ignore it.