Add configurable value for keyword split
Make the character on which keywords are temporarily joined into a string with a configuration value "metadata_keyword_separator_char".
This character is always used to join all keywords into a string and then split them apart again later.
Default this to ',' to retain existing behaviour.

Change this to an unused character e.g. chr(127) in local/config/config.inc.php to permit use of ',' in keywords:

$conf['metadata_keyword_separator_char'] = chr(127);

It is then safe to remove your favourite separator characters from metadata_keyword_separator_regex, I find this config value works
nicely if it is empty:

$conf['metadata_keyword_separator_regex'] = '';