Add last calculated cache size to Dashboard Storage chart
Size of cached files with all derivates and representatives may be
as large as all image files, actually doubling the needed storage
size. So it may be nice to see it represented in the storage
chart, if one remembers to occasionally recalculate it under
Tools -> Maintenance..

To not have this mis-representation active for non-aware users,
make it a configuration switch defaulting to false.