fixes #1986 remove document on click
the document on click gets all click and triggers more clicks. This has been removed.

Too hide a dropdown when we open another, now is specific to clicks on the filters or the filter button. We remove the class and hide the dorpdown if we have clicked on another filter