(cp 49dadcecb) fixes #2136 search for matching image_ids instead of tag_ids
Introduce an additional SQL query to find image_ids associated to matching tags instead of using directly the list of matching tags.

Same for matching albums.

Also, change the displayed list of matching tags/albums: do not take into account the OR/AND on the search by word. Do as if the search was always OR.